+++ src/Symfony/Component/Console/Tests/Helper/TableTest.php
@@ -1720,7 +1720,7 @@ EOTXT
 |-------------------------|
 |   ISBN: 9971-5-0210-0   |
 |  Title: A Tale          |
-| of Two Cities           |
+|         of Two Cities   |
 | Author: Charles Dickens |
 |  Price: 139.25          |
 +-------------------------+