+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+6.4
+---
+
+ * Add `max_retries` option to `RetryableHttpClient` to adjust the retry logic on a per request level
+
 6.3
 ---